requirements.txt
aiogram
python-dotenv
fastapi
jinja2
uvicorn 